refactor(merge-tool): Create separate helpers file for merge route
